+++ android/lib/model/src/main/kotlin/net/mullvad/mullvadvpn/lib/model/DeleteAccountError.kt
@@ -1,0 +1,5 @@
+package net.mullvad.mullvadvpn.lib.model
+
+sealed interface DeleteAccountError {
+    data class Unknown(val t: Throwable) : DeleteAccountError
+}